avocado.core: job replay support
job replay support using the run option "--replay "
implements the job replay, retrieving the original job urls, multiplex and
configuration.

The option "--replay-test-status " is fully functional,
allowing users to replay only the variants that faced a given status in
the original job.

To ignore some original job information, the option
"--replay-job-ignore " can be used.

Notice if can inform the urls and/or the multiplex using the command line.
The options informed in the command line will be used, making the replay job ignore
that information from the source job.

Signed-off-by: Amador Pahim <apahim@redhat.com>