selftests/test_interrupt.py: Handle timing problems
In InterruptTest.test_well_behaved, sometimes it
takes a bit extra time for avocado to end, causing
a psutil.NoSuchProcess exception by the time the test
tries to access .cmdline() in the Process() object.

That condition means the process is already out of
the process list, which means the test passed. Let's
handle the NoSuchProcess exception accordingly.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>