avocado.plugins.replay Limit the status list for --replay-test-status
Currently replay is using status.mapping as the list of valid status
for the --replay-test-status option. But status.mapping has some internal
status that users should not have to deal with.

This patch creates a user facing status which is a subset with only
the options relevant for the users and makes the job replay feature to take
advantage of it.

Signed-off-by: Amador Pahim <apahim@redhat.com>