Test: turn datadir into a property
The reason being that `datadir` can then be properly documented as
first class citizens so that Test class users (test writers) can rely
on it.

Other reason is that that it can be overloaded by inherited classes
such as `SimpleTest`, without the need to (re-)set other attributes
that depend on it.

Signed-off-by: Cleber Rosa <crosa@redhat.com>