avocado.utils.software_manager: Use sudo if command requires root
Instead of assuming we have administrative privileges to
run package manager related commands, use sudo in commands
that do require root. That assumes that either:

1) You're actually running as root
2) You're running as a user that has sudo configuration
   such that no password will be prompted to execute that
   command

A place where we need such a provision is when running on
Amazon AWS instances (we can't run as root, but the base
user of the image has the appropriate sudo configuration).

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>