avocado.utils.service: Use sudo if a command requires root
Instead of assuming we have administrative privileges to
run service manager related commands, use sudo in commands
that do require root. We now assume you're running the
test as a user that has sudo configuration such that no
password will be prompted to execute that command [1].

A place where we need such a provision is when running on
Amazon AWS instances (we can't run as root, but the base
user of the image has the appropriate sudo configuration).

[1] Running as root means that sudo won't be necessary,
but figuring that out is responsibility of the process
APIs, that were modified to take a sudo parameter and
figure things out based on that.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>