avocado.utils.process: Add 'sudo' parameter to run and system APIs
Some commands do require admin privileges to run, but we
can't use the 'root' user in some testing scenarios [1].

So let's add a 'sudo' parameter to process.run(), process.system()
and process.system_output() interfaces, that will append
a sudo command to the original command string, causing it
to be executed under sudo. The basic assumption here is
that the user executing the tests was added to the sudoers
file in such a way that it can run administrative commands
without a password.

[1] One specific environment where we can't use the root
user is when avocado runs on an AWS (Amazon Web Services)
instance.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>