Test: allow datadir to be optional
Some Test types are not backed directly by files, and thus do
not have a datadir. One example is Avocado-VT's VirtTest class.
Because of that, and because other custom Test types may adhere
to the same logic of not having a datadir, let's make Avocado's
base test class more robust against a `None` datadir.

Signed-off-by: Cleber Rosa <crosa@redhat.com>