Test: only publish *_BASEDIR and *_DATADIR env vars when they're set
IMHO it makes more sense to only publish these variables when they
actually contain meaningful value. Right now it *may* contain "None"
as this is maybe the value of the `basedir` and `datadir` attributes
when the test is not backed by a file.

Also, it's a lot more common, say, for a shell script to check if
a variable is set (if [ -n $AVOCADO_TEST_DATA_DIR ] ...) then to
check if it's not equal to "None".

Signed-off-by: Cleber Rosa <crosa@redhat.com>