SimpleTest: use filename property instead of path
Test and SimpleTest now have a filename property, so no need to
keep the same information in the path attribute.

Also, path is being abused by ExternalRunnerTest. Let's implement
a proper filename attribute (one that returns None, since it may not
be backed by a file), and use a custom attribute to hold the command
name that will be executed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>