[RFC] avocado.core.remoter: Turn fabric errors into CmdError exceptions
This one is a bit controversial, so I'm calling it
an RFC. It doesn't need to be merged, but hear me out:

Due to the Paramiko bug mentioned, when you execute
a lot of commands on a lot of different ssh connections,
we're bound to hit that bug sooner or later. Having
test writers to remember those 3 exceptions:

1) AuthenticationError
2) fabric.network.NetworkError
3) CmdError

Seems like a lot of overhead just to keep the concepts
entirely pure. I believe we have enough justification
to turn 1) and 2) into CmdErrors, with a special return
code -300 (I just picked a number here). This gets the
added benefit of making ignore_status=True to ignore this.

So when running probing commands on tests, that can be
simply retried and its return code ignored, working around
the Paramiko bug.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>