avocado.core.remoter: Don't hang forever on authentication problems
After some research, I found out that Paramiko (the low
level python SSH library) has bugs in key management [1],
causing some strange authentication problems from time
to time [2]. When an authentication fails in fabric by default,
it prompts the user for a password.

This means a thread executing remote commands will hang
forever. We should avoid this, so configure fabric to
throw an AuthenticationError when this happens instead
of prompting the user for a password, and handle that
error properly in our runner functions.

[1] https://github.com/paramiko/paramiko/issues/387
[2] The bug above has more details, but due to the
    way keys are handled, Paramiko gets confused and
    thinks that it needed a password from the user,
    when in actuality all key information was correct.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>