avocado.core.remoter.Remote: Support multiple instances
Since we have a fairly limited used of fabric in the vm
and remote plugins (one single host per test), we did not
notice an important limitation of the remoter code: It does
not support multiple instances of it, due to the fabric env
singleton nature. Having to connect to more than one system
with different authentication settings will cause failures
due to the fact that fabric was using old authentication
settings.

So let's use fabric execute function to make safer run
and send/receive file methods.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>