avocado.utils.process: Make "SIGCONT" optional in kill_process_tree
The "kill_process_tree" executes desired kill and then it sends
SIGCONT to the pid. In some cases (like sending SIGSTOP) this is
undesirable, but on a lot of places the code assumes it sends it.
So let's add "send_sigcont=True", which allows one to disable it
in his particular case.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>