avocado.core.runner: Use Test.timeout to obtain timeout
Currently the timeout is obtained from Test.params in runner
when early_status is obtained. This might give invalid results as
some tests override the `self.params` with other params (avocado-vt).
This patch sets the `self.timeout` value instead and uses it in
runner to avoid such problems. One can modify this value in
Test.__init__ so it should be flexible enough.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>