selftests: implement "make selfcheck" target and do it on Travis
This implements a "make selfcheck" that developers can use to run
the Avocado selftests in itself. It also keeps the old "make check"
(for a while) so that we can compare to the old approach if any
regressions occur in the near future.

Then, it switches the default runner on Travis from the simple unittest
wrapper (selftests/run) to Avocado itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>