selftests/unit/test_utils_linux_modules.py: make test a non static method
One of the unittests, namely `test_parse_lsmod_is_empty`, is implemented
as a static method, which should be fine because it doesn't use any class
attribute. But, there are a couple of problems.

First, it looks like the Python unittest library, while able to "discover"
such tests (implemented as static methods):

   $ python test_utils_linux_modules.py -v
   test_parse_lsmod (__main__.TestLsmod) ... ok
   test_parse_lsmod_is_empty (__main__.TestLsmod) ... ok
   test_parse_lsmod_no_submodules (__main__.TestLsmod) ... ok
   test_parse_lsmod_single_submodules (__main__.TestLsmod) ... ok

   ----------------------------------------------------------------------
   Ran 4 tests in 0.001s

   OK

It can not deal with it if its name is given explicitly. Example:

   $ python -m unittest test_utils_linux_modules.TestLsmod.test_parse_lsmod_is_empty
   Traceback (most recent call last):
     File "/usr/lib64/python2.7/runpy.py", line 162, in _run_module_as_main
       "__main__", fname, loader, pkg_name)
     File "/usr/lib64/python2.7/runpy.py", line 72, in _run_code
       exec code in run_globals
     File "/usr/lib64/python2.7/unittest/__main__.py", line 12, in <module>
       main(module=None)
     File "/usr/lib64/python2.7/unittest/main.py", line 94, in __init__
       self.parseArgs(argv)
     File "/usr/lib64/python2.7/unittest/main.py", line 149, in parseArgs
       self.createTests()
     File "/usr/lib64/python2.7/unittest/main.py", line 158, in createTests
       self.module)
     File "/usr/lib64/python2.7/unittest/loader.py", line 130, in loadTestsFromNames
       suites = [self.loadTestsFromName(name, module) for name in names]
     File "/usr/lib64/python2.7/unittest/loader.py", line 122, in loadTestsFromName
       (obj, test))
   TypeError: calling <function test_parse_lsmod_is_empty at 0x7fb3f58129b0> returned None, not a test

While a bug should probably be filed on upstream CPython, we can't neither
should wait on a possible fix.

Second, it's much more common, and arguably more Python, to have non-static
methods as tests, so that unittest.TestCase methods can be called, such as
"assertEquals" instead of "assert foo == bar".

So, let's change this method to be a non-static one.

Signed-off-by: Cleber Rosa <crosa@redhat.com>