TestResultProxy: remove magical __getattr__ method
This implementation of __getattr__ has allowed users of the test
result proxy, most importantly the result test runner, to ask for
attributes, including important methods, that would only be available
on the RemoteTestResult implementation.

Things have been working more or less by pure chance, based on the
lucky fact that the RemoteTestResult would be at the right place on
the proxy `output_plugins`.

Let's forbid users of the test result proxy to look for attributes
that are not part of the test result "contract" by removing this
magical (and rather evil) method.

Signed-off-by: Cleber Rosa <crosa@redhat.com>