Remote/VM: move functionality from result to runners
The functionality to run tests on remote machines and on VMs are
implemented using a mixture of TestRunner and TestResult classes.

In all fairness, it looks like the TestResult classes were abused
because non-standard attributes were used as a foundation for that
functionality. Now that `setup` and `tear_down` have been added
to the TestRunner interface, let's move those custom features from
the TestResult classes to the TestRunner ones.

Signed-off-by: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>