[WIP] selftests/unit/test_remote.py: adapt to changes from result to runner
This is WIP fix of the last unit test that needs to be touched with
the changes from remote/vm result to runner classes.

It changes the names attribute names from "remote" to "runner" for
the sake of clarity, and adds more necessary mocking (such as the
View and Job) that are necessary for the runner.

Unfortunately, it looks like even more mocking is necessary to
prevent the (real) remote from kicking in, as running that test
currently brings name resolution errors.

Signed-off-by: Cleber Rosa <crosa@redhat.com>