avocado/core/job.py: add formal preprocessing of urls
The parsed (and possibly processed) list of urls given on the command
line is quite useful to other implementation of test runners (see
the upcoming changes on the remote/vm runners). So, let's add a proper
interface that can be used to customize/sanitize the urls given on
the command line, and keep that as a reference in the `urls` attribute.

Signed-off-by: Cleber Rosa <crosa@redhat.com>