avocado.core.html: Add extensive debug on UnicodeDecodeError
Some users are getting UnicodeDecodeError, but we're unable to reproduce
it. This commit lists the full template and repr() of the failure so we
can at least see what was rendered and identify the source.

No one should actually see this message and we should remove it once we
identify and fix this problem, but without such extensive debug we can't
do anything about it.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>