avocado.plugins.run Make validated timeouts reusable
Saving args for future use with replay, the job_timeout is recorded after the
validation, in integer format. Currently we cannot reload this value
since integers are not expected for job_timeoout. Also, when no
job_timeout is provided, 0 is attributed to job_timeout, so 0 is a valid
timeout.

This patch makes run() to accept validated timeouts in integer format,
making possible to reuse them in a job replay.

Signed-off-by: Amador Pahim <apahim@redhat.com>