avocado.plugins.run Accept 0 and positive integers for job_timeout.
Saving args to further replay, the job_timeout is recorded after the
validation, in integer format. Currently we cannot reload this value
since integers are not expected for job_timeoout. Also, when no
job_timeout is provided, 0 is attributed to job_timeout, so 0 is a valid
timeout.

This patch makes run() to accept 0 or any positive integer as job_timeout,
making possible to reuse the already validated information.

Signed-off-by: Amador Pahim <apahim@redhat.com>