avocado: Refactor logging initialization
This patch changes the way logging is initialized in avocado app:

1. stdout/stderr/logging is replaced with in-memory handlers
2. when arg parsing is done, logging is reconfigured accordingly to
   setting and re-logs stored messages (when streams are enabled)
3. when job starts and "test" stream is enabled, "" and "avocado.test"
   streams are enabled
4. when job finishes and "test" stream enabled the additional outputs,
   the additional outputs are disabled again.

The 3 and 4 is necessarily as the "" logger contains all logs including
avocado initialization (stevedore, PIL, ...). This way the "test" stream
still outputs everything, but only during the job-execution and not
during avocado initialization and cleanup. (one can use "early" log to
see those).

This patch keeps the old means to enable logs (--show-job-log, --silent),
but it also adds "--log STREAMS" argument, which should be used instead.
Note that "--show-job-log" keeps the enabled streams so now you get booth
"avocado.app" and "avocado.test" in the output. While testing this seemed
easier to follow as "avocado.app" output is quite short.

Last but not least it keeps "avocado.app" output in stdout (and stderr),
but it moves all other outputs to stderr instead. This is IMO better
handling as one can distinguish between debug and "additional" output.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>