avocado.core.test: Ignore unknown arguments in TimeOutSkipTest
The TimeOutSkipTest (and heirs) are used to substitute other test
classes. Custom classes might supply additional arguments, which
TimeOutSkipTest does not handle.

This patch simply picks the supported arguments and ignores the
remaining ones.

Note that positional arguments are also used. When the order is kept, it
will work fine. But if the custom class reorders the positional
arguments things are going to be messy.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>