Test: remove stdout_log and stderr_log attributes
These two attributes are simply utility logger instances that test
writers can use to write "directly" to the test process' STDOUT and
STDERR.

I agree that having those ready to use inside a test is a nice to
have, but I don't think it justifies the polution to the test class
namespace.

Let's remove those and document how the user can use Python's standard
logging API to retrieve the logger instances that Avocado provides to
tests. This will continue to exist and be supported.

Signed-off-by: Cleber Rosa <crosa@redhat.com>