Test: remove empty implementations of setUp and tearDown
The Avocado Test class inherits from unittest.TestCase, and for that
matter it already contains implementations of setUp and tearDown.

It could be argued that the point is to better document those, but
then the same logic should be applied to all other unittest.TestCase
methods.

So, given that argument, let's remove those.

Signed-off-by: Cleber Rosa <crosa@redhat.com>