avocado.core.job: Avoid test failure on latest link failure
When avocado fails to update the link to the latest results, the whole
job is aborted. This is unfortunate and after a discussion we considered
this as a "soft" issue.

With this patch we only log an information, that the latest link was not
updated into the job-log and proceed to testing.

This commit also extends the link creation to be
concurrent-execution-safe by creating `latest-$PID` symlink first and
then renames the symlink to `latest`. The rename is atomic operation and
does not suffer the delete->create problem, where other process creates
the file before this one.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>