Test: increment to tagged name not applied on "unsafe" names
Depending on the name of the test, it may need a "safe" tagged name,
that is, one that can be turned into a filesystem path. Most, if not all,
cases apply to SIMPLE tests, that have slashes in the test names, such
as "/bin/true".

The test class code needs to check if the "safe" test name directory
exists, and apply the incremental numeric tag to it.

Also, an explicit check that raises a test setup failure was added, so
that Avocado won't silently reuse an existing test log dir.

Reference: https://trello.com/c/nPrrRv5N
Signed-off-by: Cleber Rosa <crosa@redhat.com>