avocado.core.output: Wrap the stdout/stderr handling in a class
Currently we use global variables and store combined stdout/stderr
streams into one in-memory object. This suits for ordering purposes, but
does not keep the information of what stream was the message written to.
This commit creates a wrapping class, which uses the shared storage
(list) to store messages altogether with the stream they were used with.

Additionally this class provides some basic handling of streams and we
can avoid using global variables.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>