avocado.core.app: Remove the early_logging hack
There was a hack to enable default streams when avocado failed to parse
arguments. This was necessary as --silent used to disable all output and
the user would have no idea of what went wrong. This behavior changed
and we always enable stderr leaving the user at least something to start
with. So instead of trying to be smart let's not surprise our users and
be consistent all the time.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>