avocado.core.parser: Split the argparse error message to stdout/stderr
Currently the custom argparse.error writes everything directly to
stderr. This patch replaces the direct write to suitable logging stream.

The bigger change is it splits the message and outputs the "--help"
output to stdout and puts only the error message into the stderr.

This way the output is more consistent and on colored term more
readable.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>