avocado.core.test: Initialize logdir rather than creating the name
The logdir depends on the safe-tagged-name, which is initialized in
`_get_tagged_name` function, which is checking if the safe-tagged-name
exists and reports the unsafe tagged name only to construct safe-tagged-
name few lines after and again check if the dir exists.

Instead of this this patch looks for unused safe-tagged-name and
immediately creates it and reports booth, unsafe name + the already
created logdir.

Additionally this adds adds limit on filename length and limits the
number of tries to 9999, to avoid infinite hangs.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>