avocado.core.test: Avoid None status on early exception
When avocado fails early in the process (for example in
_setup_environment_variables) it crashes, because the test status is
None. Let's define it as "ERROR" instead.

I'd much rather prefer "NOSTATUS", which is supported by avocado, but we
defined it as not-user-facing status, so results plugins can't handle
it. I'd be delighted if this requirement is dropped and we can properly
report badly initialized test.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>