avocado.utils.astring: Be more lenient regarding utf-8 chars
People tend to mix unicode, strings and encoded strings. This patch adds
"avocado.utils.astring.string_safe_encode" function, which tries to
allow all of the listed inputs and it assumes "utf-8" is in use for
encoded strings.

This function is then used in the tabular_output to avoid
"UnicodeDecodeErrors" for all listed input types.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>