Plugins: Introduce Pre and Post Job Execution Plugins
These new plugin types, registered under the 'avocado.plugins.job.pre'
and 'avocado.plugins.job.post' entry point namespaces, allow for any
custom action to be executed before and/or after the job execution.

The plugin interface requires the `run` method to be implemented.
This method will receive the current Job instance as an argument,
which allows the plugin to fetch any job information or even to tweak
as it may seem fit.

Signed-off-by: Cleber Rosa <crosa@redhat.com>