Plugin load failures: add error log utility function
With the expansion of the types of plugins, and some of them only
used in specific circurstances, they are loaded at different times.

When a plugin fails to load, wherever it happens, Avocado should log
an error. Let's extract the function that does that from the
Application and make it a utility function to be reused elsewhere.

Signed-off-by: Cleber Rosa <crosa@redhat.com>