avocado.core.loader fix crash on loader exception
When an invalid python code is present in a test class, avocado crashes
badly. This patch adds a TestLoaderError class to be used in those cases,
so we use it to fake the invalid test class, fail the test with ERROR estatus
and report the exception in job log.

Reference: https://trello.com/c/3zDIjTuY
Signed-off-by: Amador Pahim <apahim@redhat.com>