avocado.core.xunit fix XUnit results fail XSD validation
The xunit plugin claims to be compatible with jenkins-ci junit-4.xsd,
but it does not pass on the validation dur to the wrong key name 'skip'.

This patch changes the key name to 'skipped' and adds a unittest to validate
the xunit output against the junit-4.xsd schema.

Reference: https://trello.com/c/ZdEb0Tbl
Signed-off-by: Amador Pahim <apahim@redhat.com>