avocado.core.result: Remove unused API
The TestResult baseclass keeps the count of executed tests and
additionally on test end it calls function associated with the finish
status. This is not used by any plugin as they usually don't care. The
only one which could benefit from it is xunit, but it uses simplified
version. I think it's time to say good bye to this and simplify the
TestResult class (along with minor speedup)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>