avocado.core.test: Check for basename length, not full path
only the basename of the `self.filename` has to be shorter than
maxsize-5, not the full path.

Also the IOError in selftest is rather misleading as it relates to
missing source file. Let's just check the path is correctly set and
additionally verify `_record_reference_*` is not crashing when filename
not associated.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>