[Do Not Merge] Create the option to reject unknown hosts
When Avocado runs tests on remote machines, based on SSH, it always
accepts the host SSH key. This could allow credentials to be stolen
if the remote host name or address is spoofed.

This patch adds to Avocado the configuration option to enable/disable
the reject_unknown_hosts option, so user can choose between safe or
flexible.

  !!! WARNING !!!
  There is an issue with paramiko that makes the fabric reject_unknown_hosts
  option to fail:

    https://github.com/fabric/fabric/issues/85

  This PR aims to prepare Avocado so we can merge it when Paramiko/Fabric get
  fixed.

Reference: https://trello.com/c/oqyAvZBq
Signed-off-by: Amador Pahim <apahim@redhat.com>