avocado.core.remoter remote actions should fail on prompt
Avocado is not interactive. Remote can prompt for interaction for
a number of reasons: unknown host key, password request and so on.

This patch makes Avocado to fail when interaction for remote access is
needed, forwarding the correspondent Fabric's message for the user.

Reference: https://trello.com/c/GEa8BY2j
Signed-off-by: Amador Pahim <apahim@redhat.com>