Travis: improve cleanup of per-commit environment
When running the per-commit checks on CI, the previous environment is
not as clean as it should be.

One example is when a given work introduces a new plugin. First the
HEAD of the branch is tested, which includes the installation of the
new plugin.  Then, the per-commit checks would return to the first
commit in the branch, which doesn't contain the plugin, but the
environment still knows about it.  This is a matter of forcing the
clean up of setuptools entrypoints (or any setuptools generated info
for that matter).

Signed-off-by: Cleber Rosa <crosa@redhat.com>