Plugins: Introduce Pre and Post Job Execution Plugins
These new plugin interfaces, to be registered under the
'avocado.plugins.job.prepost' namespace, allow for any custom action
to be executed before and/or after the job execution.

The JobPre plugin interface requires the `pre` method to be
implemented, while the JobPost plugin interface requires the `post`
method to be implemented.  The separation of the interfaces allows for
mandatory implementation of the given interfaces, which was the whole
point of using Abstract base classes, instead of having empty default
implementations.

One single class may implement either one, or both of those
interfaces.  Both `pre` and `post` methods receive the current Job
instance as an argument, which allows the plugin to fetch any job
information or even to tweak as it may seem fit.

Signed-off-by: Cleber Rosa <crosa@redhat.com>