avocado: Rename the `args.default_multiplex_tree`
The `args.default_multiplex_tree` stands for default avocado params,
which currently supports only `multiplexer` implementation. This might
change in the future, but this variable could be used by other backends
too. So let's rename it now to `default_avocado_params`.

Additionally it's only set when the system is multiplexer-enabled, but
it has nothing to do with the multiplexer (directly). Multiplexer is
only one of the sources of params. So let's initialize it always. This
avoids a bug in `avocado-virt`, which uses this default params to set
it's params and it does that via this generic API (not dependent on the
multiplexer at all).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>