avocado: Implement serialized test ids
This commit implements the serialized test ids described in the
Introduce proper test IDs RFC.

Doing so is not essential as:

1. Files are initialized inside Test, there for we need to pass the name
to it.
2. Multiplexer does not report variant tags
3. Multiplexer/params are always involved
4. The runner is not aware of how many tests it will be executing

To solve this the Job now forwards the number of tests to runner, which
computes the number of digits from it. Then before running the test it
it wraps the test name in TestName object along with the test uid
(serialized id for now), the original name and the variant tag.

Currently each avocado execution is multiplexed as we support so called
default_avocado_params (used by avocado-virt to pass default params and
--mux-inject to inject the params provided on the cmdline). This is
always present and currently there is no way to disable it. It is
actually not strictly bound to multiplexer and should become an
independent params API soon, but not yet.

As a result I'm extending the interface of yielding the variants of
yielding the variant + the variant ID (currently serialized variant id,
but it can be anything).

The TestName class supports additional methods. The most commonly used
here are __str__, which returns the serialized-test-id (used in test
results) and the str_filesystem, which returns filesystem friendly name
according the RFC (shrinks the name, then variant or raises exception).

Things I'm not happy about:

* every test contains a variant tag (avocado always multiplexes tests)
* passing the no_tests to runner as a separate argument (IMO we should
create a TestSuite class, which stores those informations, but it's
not necessarily now)
* The whole "Test.tag" subsystem is still in place, but not used
  anymore. I'm going to remove it in the next commit to avoid too
  complex commit.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>