avocado.core.remoter fail early when host is not in known_hosts
when the user provides --remote-password and the host is not found
in known_hosts, it raises fabric.network.NetworkError. We catch this
exception in remoter as usually it means the connection timed out
and if we have enough time we try it again.

This patch checks if 'not found in known_hosts' is present in the
exception message breaks the loop to retry, re-raising the exception.

Signed-off-by: Amador Pahim <apahim@redhat.com>